% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_simulate.R
\name{fs_simulate}
\alias{fs_simulate}
\title{Produce parameter simulation of Fraser-Suzuki function}
\usage{
fs_simulate(output_folder)
}
\arguments{
\item{output_folder}{file path for output plot}
}
\value{
saved simulated Fraser-Suzuki plot
}
\description{
Produce parameter simulation of Fraser-Suzuki function
}
