% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_amabilis.R
\name{j_amabilis}
\alias{j_amabilis}
\title{J. amabilis individual curves for TGA explanation figure}
\usage{
j_amabilis(species_code, output_folder)
}
\arguments{
\item{output_folder}{output folder}

\item{list_item}{split species row as list item}

\item{species_data}{species detail file with growth form and full species name info}
}
\value{
saved deconvolved plot and list of parameters and weights
}
\description{
J. amabilis individual curves for TGA explanation figure
}
