% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_leco_traits}
\alias{load_leco_traits}
\title{Prepare C and N trait data}
\usage{
load_leco_traits(leco_data)
}
\arguments{
\item{leco_data}{raw leco data}
}
\value{
prepared leco data
}
\description{
Prepare C and N trait data
}
