% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_les_traits}
\alias{load_les_traits}
\title{Prepare LES traits}
\usage{
load_les_traits(trait_data, species_data)
}
\arguments{
\item{trait_data}{file path for raw trait data}

\item{species_data}{file with species data}
}
\value{
prepared trait data
}
\description{
Prepare LES traits
}
