% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_species.R
\name{load_species}
\alias{load_species}
\title{Prepare species data}
\usage{
load_species(species_data)
}
\arguments{
\item{species_data}{raw species data}
}
\value{
prepared species data
}
\description{
Prepare species data
}
