% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tga_traits.R
\name{load_tga_traits}
\alias{load_tga_traits}
\title{Prepare weights}
\usage{
load_tga_traits(species_deconvoluted_list)
}
\arguments{
\item{species_deconvoluted_list}{deconvoluted species data}
}
\value{
weights
}
\description{
Prepare weights
}
