% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits_log.R
\name{log_traits}
\alias{log_traits}
\title{Produce matrix of logged traits}
\usage{
log_traits(t_mean)
}
\arguments{
\item{t_mean}{mean traits dataframe}
}
\value{
matrix of logged traits
}
\description{
Produce matrix of logged traits
}
