% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_plot.R
\name{pair_plot}
\alias{pair_plot}
\title{Produce pair plot of traits}
\usage{
pair_plot(df, output_file)
}
\arguments{
\item{df}{dataframe of traits}

\item{output_file}{file path for output plot}
}
\value{
saved pair plot
}
\description{
Produce pair plot of traits
}
