% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_gf.R
\name{pca_gf}
\alias{pca_gf}
\title{Produce pca plot and loadings table}
\usage{
pca_gf(df, species_data, output_folder)
}
\arguments{
\item{df}{dataframe of traits}

\item{output_folder}{folder path for pca plot}
}
\value{
saved pca plot and loadings table
}
\description{
Produce pca plot and loadings table
}
