% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_loadings.R
\name{pca_loadings}
\alias{pca_loadings}
\title{Produce pca plot and loadings table}
\usage{
pca_loadings(prin, output_file)
}
\arguments{
\item{prin}{pca output}

\item{output_file}{file path for loadings table}
}
\value{
saved loadings table
}
\description{
Produce pca plot and loadings table
}
