% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{phylo_accessions}
\alias{phylo_accessions}
\title{Produce table of GenBank Accessions}
\usage{
phylo_accessions(genbank_accessions_file, output_file)
}
\arguments{
\item{genbank_accessions_file}{file path for accessions info}

\item{output_file}{file path to save table}
}
\value{
saved table file
}
\description{
Produce table of GenBank Accessions
}
