% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{phylo_mantel}
\alias{phylo_mantel}
\title{Produce table of estimates of Mantel test of tree distance and trait distance}
\usage{
phylo_mantel(phylo, tips, output_file)
}
\arguments{
\item{phylo}{phylogenetic tree}

\item{tips}{traits}

\item{output_file}{where to save mantel table}
}
\value{
saved mantel table results
}
\description{
Produce table of estimates of Mantel test of tree distance and trait distance
}
