% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{phylo_plot}
\alias{phylo_plot}
\title{Produce phylo plot}
\usage{
phylo_plot(phylo, tips)
}
\arguments{
\item{phylo}{phylogenetic tree data}

\item{tips}{phylogenetic traits}
}
\value{
saved pca plot
}
\description{
Produce phylo plot
}
