% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{phylo_readtree}
\alias{phylo_readtree}
\title{Read phylogenetic tree data}
\usage{
phylo_readtree(nwk_file)
}
\arguments{
\item{nwk_file}{file path for tree}
}
\value{
tree as phylogeny
}
\description{
Read phylogenetic tree data
}
