% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{phylo_traits}
\alias{phylo_traits}
\title{Add functional trait data to phylogenetic tree data}
\usage{
phylo_traits(phylo, t_mean)
}
\arguments{
\item{phylo}{phylogenetic tree object}

\item{t_mean}{mean traits dataframe}
}
\value{
phylogenetic data with traits
}
\description{
Add functional trait data to phylogenetic tree data
}
