% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeny.R
\name{phylogeny}
\alias{phylogeny}
\title{Produce phylogeny outputs}
\usage{
phylogeny(df, genbank_accessions_file, nwk_file, figure_folder, table_folder)
}
\arguments{
\item{df}{dataframe of mean trait values}

\item{genbank_accessions_file}{file path for genbank accession codes}

\item{nwk_file}{file path for .nwk file}

\item{output_folder}{folder path for plot output}
}
\value{
saved phylogeny files
}
\description{
Produce phylogeny outputs
}
