% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_leco.R
\name{prep_leco}
\alias{prep_leco}
\title{Prepare C and N trait data}
\usage{
prep_leco(leco_data)
}
\arguments{
\item{leco_data}{raw leco data}
}
\value{
prepared leco data
}
\description{
Prepare C and N trait data
}
