% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tga.R
\name{prep_tga}
\alias{prep_tga}
\title{Prepare species' TGA data}
\usage{
prep_tga(species_data, output_folder)
}
\arguments{
\item{species_data}{species detail file with growth form and full species name info}

\item{output_folder}{output folder}
}
\value{
saved deconvolved plot and weight
}
\description{
Prepare species' TGA data
}
