% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_tga.R
\name{process_raw_tga}
\alias{process_raw_tga}
\title{Load and process raw TGA data}
\usage{
process_raw_tga(raw_file)
}
\arguments{
\item{raw_file}{TGA raw data}
}
\value{
processed TGA data
}
\description{
Load and process raw TGA data
}
