% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_fraser_suzuki.R
\name{simulate_fraser_suzuki}
\alias{simulate_fraser_suzuki}
\title{Produce parameter simulation of Fraser-Suzuki function}
\usage{
simulate_fraser_suzuki(output_file)
}
\arguments{
\item{output_file}{file path for output plot}
}
\value{
saved simulated Fraser-Suzuki plot
}
\description{
Produce parameter simulation of Fraser-Suzuki function
}
