% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{single_deconvolute}
\alias{single_deconvolute}
\title{Deconvolve a single TGA curve}
\usage{
single_deconvolute(raw_file)
}
\arguments{
\item{raw_file}{TGA raw data}

\item{subfig}{sub figure label ID}

\item{output_file}{output file name}
}
\value{
saved deconvolved plot and weight
}
\description{
Deconvolve a single TGA curve
}
