% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_deconvolve.R
\name{single_deconvolve}
\alias{single_deconvolve}
\title{Deconvolve a single TGA curve}
\usage{
single_deconvolve(raw_file, subfig, output_file)
}
\arguments{
\item{raw_file}{TGA raw data}

\item{subfig}{sub figure label ID}

\item{output_file}{output file name}
}
\value{
saved deconvolved plot and weight
}
\description{
Deconvolve a single TGA curve
}
