% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_deconvolute.R
\name{species_deconvolute}
\alias{species_deconvolute}
\title{Deconvolute species' TGA data}
\usage{
species_deconvolute(list_item, species_data, output_folder)
}
\arguments{
\item{list_item}{split species row as list item}

\item{species_data}{species detail file with growth form and full species name info}

\item{output_folder}{output folder}
}
\value{
saved deconvolved plot and list of parameters and weights
}
\description{
Deconvolute species' TGA data
}
