% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tga_param_table}
\alias{tga_param_table}
\title{Produce parameter values table}
\usage{
tga_param_table(species_deconvoluted_list, species_data, output_file)
}
\arguments{
\item{species_deconvoluted_list}{list of deconvolved}

\item{species_data}{dataframe with species info}

\item{output_folder}{folder path for output plot}
}
\value{
saved parameters tables
}
\description{
Produce parameter values table
}
