% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tga_plots.R
\name{tga_plot}
\alias{tga_plot}
\title{Plot species' TGA data}
\usage{
tga_plot(species_deconvoluted_list, species_data, output_folder)
}
\arguments{
\item{species_deconvoluted_list}{deconvoluted species data list output}

\item{species_data}{species detail file with growth form and full species name info}

\item{output_folder}{output folder}
}
\value{
saved deconvolved plots
}
\description{
Plot species' TGA data
}
