% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{tga_plot_gram}
\alias{tga_plot_gram}
\title{Call TGA plot for graminoids}
\usage{
tga_plot_gram(species_deconvoluted_list, species_data, subfig, gf)
}
\arguments{
\item{species_deconvoluted_list}{deconvoluted species data list item}

\item{species_data}{species detail file with growth form and full species name info}

\item{subfig}{subfig label}

\item{gf}{growth form to plot species of}
}
\value{
saved multi-panel TGA plot
}
\description{
Call TGA plot for graminoids
}
