% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{tga_plot_others}
\alias{tga_plot_others}
\title{Call TGA plot for tree, nonvascular, and shrub}
\usage{
tga_plot_others(species_deconvoluted_list, species_data, subfigs)
}
\arguments{
\item{species_deconvoluted_list}{deconvoluted species data list item}

\item{species_data}{species detail file with growth form and full species name info}

\item{subfigs}{subfigure labels}
}
\value{
saved multi-panel TGA plot
}
\description{
Call TGA plot for tree, nonvascular, and shrub
}
