% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{tga_plot_three}
\alias{tga_plot_three}
\title{Call TGA plot for three emblem species}
\usage{
tga_plot_three(species_deconvoluted_list, species_data, species_names)
}
\arguments{
\item{species_deconvoluted_list}{deconvoluted species data list item}

\item{species_data}{species detail file with growth form and full species name info}

\item{species_names}{to plot}
}
\value{
saved multi-panel TGA plot
}
\description{
Call TGA plot for three emblem species
}
