% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tga_plots_legend.R
\name{tga_plots_legend}
\alias{tga_plots_legend}
\title{Produce TGA figure legend}
\usage{
tga_plots_legend(sample_data_file, output_file)
}
\arguments{
\item{sample_data_file}{file path to sample tga csv data for legend construction}

\item{output_file}{file path for output plot}
}
\value{
saved TGA legend
}
\description{
Produce TGA figure legend
}
