% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{tga_raw_plots}
\alias{tga_raw_plots}
\title{Deconvolute raw materials plot}
\usage{
tga_raw_plots(item_1, item_2)
}
\arguments{
\item{item_1}{first deconvolve raw object}

\item{item_2}{second deconvolve raw object}
}
\value{
saved deconvolved plot of both raw materials
}
\description{
Deconvolute raw materials plot
}
