% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tga_species_deconvolute.R
\name{tga_species_deconvolute}
\alias{tga_species_deconvolute}
\title{Prepare species' TGA data}
\usage{
tga_species_deconvolute(tga_folder, species_data)
}
\arguments{
\item{tga_folder}{folder where raw files are stored}

\item{species_data}{species detail file with growth form and full species name info}
}
\value{
saved deconvolved plots and weights
}
\description{
Prepare species' TGA data
}
