% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{tga_wrapper}
\alias{tga_wrapper}
\title{Wrapper for all species for TGA functions}
\usage{
tga_wrapper(species, function_name, ...)
}
\arguments{
\item{...}{optional arguments}

\item{list}{input data list}
}
\value{
output of internal function
}
\description{
Wrapper for all species for TGA functions
}
