% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_boxplot.R
\name{trait_boxplot}
\alias{trait_boxplot}
\title{Produce boxplot}
\usage{
trait_boxplot(df, output_folder)
}
\arguments{
\item{df}{dataframe of traits}

\item{output_folder}{file path for folder for output plot}
}
\value{
saved box plot
}
\description{
Produce boxplot
}
