% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits_combine.R
\name{traits_combine}
\alias{traits_combine}
\title{Merge trait datasets}
\usage{
traits_combine(species, trt, cn, tga)
}
\arguments{
\item{species}{species info dataframe}

\item{trt}{les trait data}

\item{cn}{carbon and nitrogen leco data}

\item{tga}{tga data}
}
\value{
merged trait datasets
}
\description{
Merge trait datasets
}
