% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{traits_log}
\alias{traits_log}
\title{Produce matrix of logged traits}
\usage{
traits_log(t_mean)
}
\arguments{
\item{t_mean}{mean traits dataframe}
}
\value{
matrix of logged traits
}
\description{
Produce matrix of logged traits
}
