% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{traits_mean_only}
\alias{traits_mean_only}
\title{Select mean of traits}
\usage{
traits_mean_only(all_traits, output_file)
}
\arguments{
\item{all_traits}{dataframe of traits}

\item{output_file}{file path for output plot}
}
\value{
mean traits only
}
\description{
Select mean of traits
}
