% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{traits_table}
\alias{traits_table}
\title{Produce output trait table}
\usage{
traits_table(traits_df, output_file)
}
\arguments{
\item{traits_df}{dataframe of traits}

\item{output_file}{file path for output plot}
}
\value{
saved trait table
}
\description{
Produce output trait table
}
