% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Produce pca plot and loadings table}
\usage{
pca(prin, df, species_data, output_file)
}
\arguments{
\item{prin}{pca data output}

\item{df}{logged traits matrix}

\item{species_data}{species dataframe so can get the abreviated species labels}

\item{output_file}{file path for pca plot}
}
\value{
saved pca plot
}
\description{
Produce pca plot and loadings table
}
