% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_gf.R
\name{pca_gf}
\alias{pca_gf}
\title{Produce pca plot and loadings table}
\usage{
pca_gf(df, figure_folder, table_folder)
}
\arguments{
\item{df}{dataframe of traits}

\item{figure_folder}{folder path for pca plot}

\item{table_folder}{folder path for pca table}
}
\value{
saved pca plot and loadings table
}
\description{
Produce pca plot and loadings table
}
