% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_les.R
\name{prep_les}
\alias{prep_les}
\title{Prepare LES traits}
\usage{
prep_les(trait_file, species_data)
}
\arguments{
\item{trait_file}{file path for raw trait data}

\item{species_data}{file with species data}
}
\value{
prepared trait data
}
\description{
Prepare LES traits
}
