% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{simulate_fraser_suzuki}
\alias{simulate_fraser_suzuki}
\title{Produce parameter simulation of Fraser-Suzuki function}
\usage{
simulate_fraser_suzuki()
}
\value{
saved simulated Fraser-Suzuki plot
}
\description{
Produce parameter simulation of Fraser-Suzuki function
}
