% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{tga_plot}
\alias{tga_plot}
\title{Plot single species' TGA data}
\usage{
tga_plot(species_code, species_deconvoluted_list, species_data,
  legend_species = TRUE)
}
\arguments{
\item{species_data}{species detail file with growth form and full species name info}

\item{deconvoluted_list_item}{deconvoluted species data list item}

\item{output_folder}{output folder}
}
\value{
saved deconvolved plot
}
\description{
Plot single species' TGA data
}
