% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tga_theory_plots.R
\name{tga_theory_plots}
\alias{tga_theory_plots}
\title{Individual curves for TGA theory explanation figure}
\usage{
tga_theory_plots(tga_data, output_folder)
}
\arguments{
\item{output_folder}{output folder to save plots}

\item{species_code}{species code to use in theory plot}
}
\value{
three theory plots
}
\description{
Individual curves for TGA theory explanation figure
}
